pyproject.toml
[project]
name = "transformers-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "av>=16.1.0",
    "bitsandbytes>=0.49.1",
    "dall-e>=0.1",
    "datasets>=4.4.2",
    "evaluate>=0.4.6",
    "huggingface-hub>=0.36.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "nbconvert>=7.16.6",
    "nltk>=3.9.2",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "protobuf>=6.33.4",
    "pycocotools>=2.0.11",
    "pytorch-lightning>=2.6.0",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentencepiece>=0.2.1",
    "seqeval>=1.2.2",
    "soundfile>=0.13.1",
    "torch>=2.9.1",
    "torchcodec>=0.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wandb>=0.23.1",
    "yt-dlp>=2026.3.3",
]
